xargs: Allow running interactive commands with -o
-o opens /dev/tty and connects it to the stdin of the commands that get
executed.

Example: say we want to delete some files interactively. We cannot do
that with xargs without the -o option because the `rm -i` command does
not receive an interactive stdin. But if we add `-o`, then xargs will
ensure that /dev/tty is connected as stdin and `rm -i` will show us a
prompt and wait:

```
% touch a b c
% echo a b c | xargs -o rm -i
```